--- Client/Components/AuditEntryViewer/AuditEntryViewer.razor
@@ -1,49 +1,0 @@
-@namespace FileFlows.Client.Components
-
-
-@if (Visible)
-{
-    <div class="audit-entry-viewer-wrapper">
-        <div class="background-overlay"></div>
-        <div class="audit-entry-viewer @(Maximised ? "maximised" : "")">
-
-            <ViContainer Title="@lblTitle" AlwaysShowTitle=@true Maximise=true OnMaximised=@OnMaximised>
-
-                <Head>
-                    <button class="btn" @onclick=@Close>@lblClose</button>
-                </Head>
-
-                <Body>
-                <FlowTable @ref=Table TItem="EntryViewerData" Selection="SelectionMode.None" Data=@Data TableIdentifier="AuditEntry" HideToolbar="true">
-                    <Columns>
-                        <FlowTableColumn TItem="EntryViewerData" Width="14rem" NoHeight="true">
-                            <Header>
-                                @lblField
-                            </Header>
-                            <Cell Context="item">
-                                <span>@item.Name</span>
-                            </Cell>
-                        </FlowTableColumn>
-                        <FlowTableColumn TItem="EntryViewerData" NoHeight="true">
-                            <Header>
-                                @lblValue
-                            </Header>
-                            <Cell Context="item">
-                                @if (item.Value is MarkupString mkString)
-                                {
-                                    <span style="white-space: pre">@mkString</span>
-                                }
-                                else
-                                {
-                                    <span style="white-space: pre">@item.Value.ToString()</span>
-                                }
-                            </Cell>
-                        </FlowTableColumn>
-                    </Columns>
-                </FlowTable>
-                </Body>
-            </ViContainer>
-        </div>
-
-    </div>
-}
