+++ Client/Shared/MainLayout/MainLayout.razor
@@ -46,14 +46,4 @@
     </div>
 </div>
 
-
-
-@if (Blocker != null)
-{
-    <CascadingValue Value="Blocker" IsFixed="true">
-        <AuditHistory/>
-        <AuditEntryViewer/>
-    </CascadingValue>
-}
-
 <FileFlows.Client.Components.ModalRenderer />
